requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
flatbuffers==24.3.25
grpcio==1.64.1
grpcio-tools==1.64.1
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.26.4
onnx-predict-yolov8==1.0.6
onnxruntime==1.18.0
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
protobuf==5.27.1
pyTelegramBotAPI==4.19.1
requests==2.32.3
sympy==1.12.1
telebot==0.0.5
urllib3==2.2.1
